recommendation/requirements.txt
psycopg2-binary
Flask

